+++ framework/domain/src/main/java/is/codion/framework/domain/entity/attribute/DefaultAttribute.java
@@ -131,7 +131,7 @@ final class DefaultAttribute<T> implements Attribute<T>, Serializable {
 		}
 
 		@Override
-		public boolean isNumerical() {
+		public boolean isNumeric() {
 			return Number.class.isAssignableFrom(valueClass);
 		}