+++ changelog.md
@@ -10,6 +10,7 @@ Codion Change Log
 - ComponentBuilder.font(UnaryOperator<Font>) added, which derives the font from the default component font.
 ### is.codion.framework.domain
 - DefaultEntityValidator removed, validation moved to AbstractAttributeDefinition, EntityValidator now with default method implementations.
+- Attribute.Type.isNumerical() renamed to isNumeric().
 ### is.codion.framework.db
 - DefaultSelect bug fixed, having clause not included in hashCode().
 - Select.having() added, related refactoring.